+++ src/app/device-order/ais/device-order-ais-existing-gadget/containers/device-order-ais-existing-gadget-effective-start-date-page/device-order-ais-existing-gadget-effective-start-date-page.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, OnChanges, OnDestroy, SimpleChanges } from '@angular/core';
-import { ShoppingCart, HomeService, PageLoadingService, AlertService } from 'mychannel-shared-libs';
+import { ShoppingCart, HomeService, PageLoadingService, AlertService, User, TokenService } from 'mychannel-shared-libs';
 import { WIZARD_DEVICE_ORDER_AIS } from 'src/app/device-order/constants/wizard.constant';
 import { PriceOption } from 'src/app/shared/models/price-option.model';
 import { Transaction, BillingAccount } from 'src/app/shared/models/transaction.model';
@@ -52,6 +52,7 @@ export class DeviceOrderAisExistingGadgetEffectiveStartDatePageComponent impleme
   shoppingCart: ShoppingCart;
   wizards: string[] = WIZARD_DEVICE_ORDER_AIS;
   mobileNoSelect: string;
+  user: User;
 
   priceOption: PriceOption;
   transaction: Transaction;
@@ -98,6 +99,7 @@ export class DeviceOrderAisExistingGadgetEffectiveStartDatePageComponent impleme
     private transactionService: TransactionService,
     private shoppingCartService: ShoppingCartService,
     private translateService: TranslateService,
+    private tokenService: TokenService,
     private http: HttpClient,
     private fb: FormBuilder,
     private pageLoadingService: PageLoadingService,
@@ -107,17 +109,21 @@ export class DeviceOrderAisExistingGadgetEffectiveStartDatePageComponent impleme
     this.transaction = this.transactionService.load();
     this.shoppingCart = this.shoppingCartService.getShoppingCartData();
     this.mobileNoSelect = this.shoppingCart.mobileNo;
-    this.homeService.callback = () => {
-      this.alertService.question('ต้องการยกเลิกรายการขายหรือไม่ การยกเลิก ระบบจะคืนสินค้าเข้าสต๊อคสาขาทันที', 'ตกลง', 'ยกเลิก')
-        .then((response: any) => {
-          if (response.value === true) {
-            this.returnStock().then(() => {
-              this.transactionService.remove();
-              window.location.href = '/';
-            });
-          }
-        });
-    };
+    this.user = this.tokenService.getUser();
+    if (this.transaction && this.transaction.data && this.transaction.data.order && this.transaction.data.order.soId) {
+      this.homeService.callback = () => {
+        this.alertService.question('ต้องการยกเลิกรายการขายหรือไม่ การยกเลิก ระบบจะคืนสินค้าเข้าสต๊อคสาขาทันที', 'ตกลง', 'ยกเลิก')
+          .then((response: any) => {
+            if (response.value === true) {
+              this.returnStock().then(() => {
+                this.transaction.data.order = {};
+                this.transactionService.remove();
+                window.location.href = '/';
+              });
+            }
+          });
+      };
+    }
   }
 
   ngOnInit(): void {
@@ -265,10 +271,9 @@ export class DeviceOrderAisExistingGadgetEffectiveStartDatePageComponent impleme
       const promiseAll = [];
       if (transaction.data) {
         if (transaction.data.order && transaction.data.order.soId) {
-          const order = this.http.post('/api/salesportal/device-sell/item/clear-temp-stock', {
-            location: this.priceOption.productStock.location,
+          const order = this.http.post('/api/salesportal/dt/remove-cart', {
             soId: transaction.data.order.soId,
-            transactionId: transaction.transactionId
+            userId: this.user.username
           }).toPromise().catch(() => Promise.resolve());
           promiseAll.push(order);
         }
